+++ src/at/sume/dm/model/residential_satisfaction/SocialPrestige.java
@@ -21,7 +21,7 @@ public class SocialPrestige extends ResidentialSatisfactionComponent {
 	 * @see at.sume.dm.model.residential_satisfaction.ResidentialSatisfactionComponent#calc(at.sume.dm.entities.HouseholdRow, at.sume.dm.entities.SpatialUnitRow)
 	 */
 	@Override
-	public long calc(HouseholdRow household, DwellingRow dwelling, SpatialUnitRow spatialUnit, int modelYear) {
+	public short calc(HouseholdRow household, DwellingRow dwelling, SpatialUnitRow spatialUnit, int modelYear) {
 		// Calculate household income per member
 		long hhIncome = household.getYearlyIncomePerMemberWeighted();
 		// Calculate average household income per member
@@ -31,7 +31,7 @@ public class SocialPrestige extends ResidentialSatisfactionComponent {
 		// TODO: individual threshold for this comparison?!?
 		if ((hhIncome > avgIncome) && (hhIncome > 0)) {
 			// satisfaction = percentage of avg. income to household income
-			return (avgIncome * 1000) / hhIncome;
+			return (short) ((avgIncome * 1000) / hhIncome);
 		} else {
 			// 100% satisfaction here
 			return 1000;
